fixed create question, get all question by group name
